Add support to parse char argument
The change adds support to parse a char argument from a format string,
e.g.`defmt::info("x={:char}", 'y')`. For now the char is converted and
written as `u32` without any compression or using UTF8.

This is mainly to have tests and format in place, it can be iterated on.