Use primitive formatting for primitive references
Closes #163

A single recursive call in `as_native_type` is sufficient
to unpack the type from the reference.
Formatting functions (`fmt.i32()` etc.) are already taking
a reference, so this does not need to be changed.

Tests are adjusted, so that they pass.
This introduced change is not explicitly tested though.