write! macro: remove runtime check
when this feature was added, it was possible to call `write!` more than once from a `Format::format`
implementation. that operation corrupted the data stream so a runtime check was added to `write!` to
prevent the data corruption by panicking on a double use of `write!`

PR #305 makes `write!` consume the `Formatter` value so the above issue is now prevented at compile
time. The runtime check is no longer necessary (failure case is unreachable) so we remove it