Update mod.rs
Make the error message more meaningful. 

Currently it tells you probe-run supports defmt X while the firmware supports defmt Y, which is reasonable. But the substitution then tells you to install a version of probe-run that supports X (the version you're already using), which is contradictory, since Y (firmware supported version) is the real version you need to install for.