Perform typechecking on disabled log statements
Without this change, the body of log statements is completely removed
from the code if the log statement is disabled by the `DEFMT_LOG` env
variable. This means that even invalid log statements will compile.

This change ensures that the code logging code is still visible to the
compiler, but inside an `if false { ... }` block, so it can easily be
thrown away by the compiler after type checking.