add `impl Format for std::sync::Mutex`
this introduces a new optional feature `std` and an initial `Format`
implementation just for `Mutex`. the hope is that over time more and
more `Format` impls for `std` can be added for those who need it.

the trigger for this is that i broke the `embassy-sync` build for `std`
by adding a `Format` derive to something which was using `Mutex`
internally and it wasn't covered by CI there (which by now is
rectified). in the ensuing conversation on Matrix it was stated that it
might be nice to have `defmt` impls for `std` as well, even if only very
few users would exist for this (i'm not one of them, which is how i
ended up breaking CI).

part of #979